requirements.txt
python-dotenv==1.1.0
pandas==1.3.5
matplotlib==3.2.2
notebook==7.4.0